Remove old exports
