+++ packages/core-utils/src/coders/index.ts
@@ -1,3 +1,2 @@
 export * from './ecdsa-coder'
-export * from './types'
 export * from './sequencer-batch'
