+++ packages/core-utils/src/index.ts
@@ -3,7 +3,6 @@ import { ethers } from 'ethers'
 
 export const abi = new ethers.utils.AbiCoder()
 
-export * from './transport'
 export * from './constants'
 export * from './log'
 export * from './hex-strings'
