--- packages/core-utils/src/transport/client/index.ts
@@ -1,4 +1,0 @@
-export * from './axios-http-client'
-export * from './json-rpc-client'
-export * from './json-rpc-http-adapter'
-export { SimpleClient } from './simple-client'
