--- packages/core-utils/src/transport/server/simple-server.ts
@@ -1,1 +1,0 @@
-export { JsonRpcServer as SimpleServer } from './json-rpc-server'
