+++ packages/core-utils/src/types/coder-types.ts
@@ -1,0 +1,4 @@
+export interface Coder {
+  encode: Function
+  decode: Function
+}
