+++ packages/core-utils/src/types/index.ts
@@ -1,3 +1,2 @@
-export * from './log.types'
-export * from './transport.types'
-export * from './type-guards'
+export * from './coder-types'
+export * from './eth-types'
