--- packages/core-utils/src/types/log.types.ts
@@ -1,6 +1,0 @@
-export interface Logger {
-  debug: (...args: any[]) => any
-  info: (...args: any[]) => any
-  warn: (...args: any[]) => any
-  error: (...args: any[]) => any
-}
